+++ pkgs/test_api/lib/src/backend/configuration/timeout.dart
@@ -2,8 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:string_scanner/string_scanner.dart';
 import 'package:meta/meta_meta.dart';
+import 'package:string_scanner/string_scanner.dart';
 
 /// A regular expression that matches text until a letter or whitespace.
 ///